refactor: implement better eks client session handler
